+++ src/Crypto/Crypto.service.ts
@@ -4,8 +4,8 @@ import { Report } from 'src/DbRepository/Report/Report.entity';
 import { Repository } from 'typeorm';
 import { getCurrencyData } from './utils';
 import { ReportCurrency } from 'src/DbRepository/ReportCurrency/ReportCurrency.entity';
-import { ReportRepositoryHandler } from 'src/classes/ReportRepository/ReportRepository';
-import { ReportCurrencyRepositoryHandler } from 'src/classes/ReportCurrency/ReportCurrencyHandler';
+import { ReportRepositoryHandler } from 'src/classes/ReportRepositoryHandler/ReportRepositoryHandler';
+import { ReportCurrencyRepositoryHandler } from 'src/classes/ReportCurrencyHandler/ReportCurrencyHandler';
 import { GetCurrencyDataInput } from './types';
 
 @Injectable()
