+++ src/classes/ReportCurrencyHandler/ReportCurrencyHandler.ts
